moved folder into root

+++ test/CoinFlip.t.sol
@@ -1,9 +1,9 @@
 pragma solidity ^0.8.10;
 
 import "ds-test/test.sol";
-import "../CoinFlip/CoinFlipHack.sol";
-import "../CoinFlip/CoinFlipFactory.sol";
-import "../Ethernaut.sol";
+import "../src/CoinFlip/CoinFlipHack.sol";
+import "../src/CoinFlip/CoinFlipFactory.sol";
+import "../src/Ethernaut.sol";
 import "./utils/vm.sol";
 
 contract CoinFlipTest is DSTest {
@@ -37,17 +37,19 @@ contract CoinFlipTest is DSTest {
         CoinFlipHack coinFlipHack = new CoinFlipHack(levelAddress);
 
         // Run the attack 10 times, iterate the block each time, function can only be called once per block
-        for (uint i = 0; i <= 10; i++) { 
+        for (uint256 i = 0; i <= 10; i++) {
             // Must be on latest version of foundry - blockhash was defaulting to 0 in earlier version of foundry resolved in this commit https://github.com/gakonst/foundry/pull/728
             vm.roll(6 + i);
-            coinFlipHack.attack();        
+            coinFlipHack.attack();
         }
 
         //////////////////////
         // LEVEL SUBMISSION //
         //////////////////////
 
-        bool levelSuccessfullyPassed = ethernaut.submitLevelInstance(payable(levelAddress));
+        bool levelSuccessfullyPassed = ethernaut.submitLevelInstance(
+            payable(levelAddress)
+        );
         vm.stopPrank();
         assert(levelSuccessfullyPassed);
     }
